can navigate through screens all the way to maps
