+++ App.js
@@ -13,7 +13,7 @@ export default class App extends React.Component {
   render() {
     return (
       // <RootStack />
-      <Map />
+      <MapScreen />
     )
   }
 }
