--- components/Map.js
@@ -1,138 +1,0 @@
-import React, { Component } from 'react'
-import { 
-	AppRegistry, 
-	StyleSheet, 
-	Dimensions, 
-	Image, 
-	View, 
-	Text,
-	StatusBar, 
-	TouchableOpacity 
-} from "react-native";
-import MapView, { Marker } from 'react-native-maps'
-import MapViewDirections from 'react-native-maps-directions'
-
-const GOOGLE_MAPS_APIKEY = 'AIzaSyDNIsEsuc8FsHQJsswUcDKUd9k3sZqzk3U'
-
-class Map extends Component {
-	state = {
-		region: {
-			latitude: 36.1699412,
-      longitude: -115.1398295,
-      latitudeDelta: 0.5,
-      longitudeDelta: 0.0421
-		}
-	}
-
-	// showCurrentLocation() {
-	// 	this.watchId = navigator.geolocation.watchPosition(
-	// 		(position) => {
-	// 			this.setState({
-	// 				latitude: position.coords.latitude,
-	// 				longitude: position.coords.longitude,
-	// 				error: null
-	// 			})
-	// 		},
-	// 		(error) => this.setState({ error: error.message }),
-	// 		{ enableHighAccuracy: true, timeout: 20000, maximumAge: 1000, distanceFilter: 10 }
-	// 	)
-	// }
-
-	componentDidMount() {
-		// this.showCurrentLocation()
-		this.watchId = navigator.geolocation.watchPosition(
-			(position) => {
-				this.setState({
-					latitude: position.coords.latitude,
-					longitude: position.coords.longitude,
-					error: null
-				})
-			},
-			(error) => this.setState({ error: error.message }),
-			{ enableHighAccuracy: true, timeout: 20000, maximumAge: 1000, distanceFilter: 10 }
-		)
-	}
-
-	componentWillUnmount() {
-		navigator.geolocation.clearWatch(this.watchId)
-	}
-
-	render() {
-
-		return (
-			<View style={styles.container}>
-				<MapView style={styles.map}
-			    initialRegion={this.state.region}
-					followsUserLocation: {true}
-					showsMyLocationButton: {true}
-					showsCompass: {true}>
-
-				  <MapView.Marker
-						coordinate={{
-							latitude: this.state.latitude,
-							longitude: this.state.longitude
-						}}
-						title={'Current location'}
-						description={'A school for gifted children.'}
-				  />
-				  <MapView.Marker
-						coordinate={{
-							latitude: 36.117388,
-    					longitude: -115.169968
-						}}
-						title={'In-N-Out Burger'}
-						description={'The food of the people.'}
-				  />
-
-				  <MapViewDirections 
-				  	origin={{
-				  		latitude: this.state.latitude,
-							longitude: this.state.longitude
-				  	}}
-				  	destination={{
-				  		latitude: 36.117388,
-    					longitude: -115.169968
-				  	}}
-				  	apikey={GOOGLE_MAPS_APIKEY}
-				  	strokeWidth={3}
-				  	strokeColor='blue'
-				  	language='en'
-				  />
-		  	</MapView>
-
-		  	<Text>Latitude: {this.state.latitude}</Text>
-		  	<Text>Latitude: {this.state.longitude}</Text>
-		  	{this.state.error ? <Text>Error: {this.state.error}</Text> : null}
-				
-				<Text>French Toast</Text>
-				<Text>Potatoes</Text>
-				<Text>French Fries</Text>
-				<Text>Spaghetti</Text>
-				<Text>French Bread</Text>
-				<Text>Cheeseburger</Text>
-		  </View>
-		)
-	}
-}
-
-export default Map
-
-const styles = StyleSheet.create({
-	container: {
-    // flex: 1,
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    bottom: 0,
-    right: 0,
-    justifyContent: 'flex-end',
-    alignItems: 'center'
-  },
-  map: {
-  	position: 'absolute',
-  	top: 20,
-  	left: 0,
-    bottom: 180,
-    right: 0
-  }
-})
